fix all msvc W3 warnings

+++ code/foundation/src/world/world_view.c
@@ -21,7 +21,7 @@ int32_t tracker_read_update(librg_world *w, librg_event *e) {
     size_t actual_length = librg_event_size_get(w, e);
     char *buffer = librg_event_buffer_get(w, e);
     world_view *view = (world_view*)librg_world_userdata_get(w);
-    
+
     entity_view data = entity_view_unpack_struct(buffer, actual_length);
     entity_view *d = entity_view_get(&view->entities, entity_id);
 #if 1
@@ -33,8 +33,8 @@ int32_t tracker_read_update(librg_world *w, librg_event *e) {
         else return 0;
     }
 #endif
-    
-    data.last_update = get_cached_time()*1000.0f;
+
+    data.last_update = (uint64_t)(get_cached_time()*1000.0f);
     data.layer_id = view->active_layer_id;
     predict_receive_update(d, &data);
     entity_view_update_or_create(&view->entities, entity_id, data);
@@ -48,7 +48,7 @@ int32_t tracker_read_create(librg_world *w, librg_event *e) {
     size_t actual_length = librg_event_size_get(w, e);
     char *buffer = librg_event_buffer_get(w, e);
     world_view *view = (world_view*)librg_world_userdata_get(w);
-    
+
     entity_view data = entity_view_unpack_struct(buffer, actual_length);
     data.ent_id = entity_id;
     data.layer_id = view->active_layer_id;
@@ -80,11 +80,11 @@ void world_view_init(world_view *view, uint32_t seed, uint64_t ent_id, uint16_t 
     view->chunk_amount = chunk_amount;
     view->dim = WORLD_BLOCK_SIZE * chunk_size * chunk_amount;
     view->size = view->dim * view->dim;
-    
+
     librg_config_chunksize_set(view->tracker, WORLD_BLOCK_SIZE * chunk_size, WORLD_BLOCK_SIZE * chunk_size, 1);
     librg_config_chunkamount_set(view->tracker, chunk_amount, chunk_amount, 1);
     librg_config_chunkoffset_set(view->tracker, LIBRG_OFFSET_BEG, LIBRG_OFFSET_BEG, LIBRG_OFFSET_BEG);
-    
+
     librg_event_set(view->tracker, LIBRG_READ_CREATE, tracker_read_create);
     librg_event_set(view->tracker, LIBRG_READ_REMOVE, tracker_read_remove);
     librg_event_set(view->tracker, LIBRG_READ_UPDATE, tracker_read_update);
